Added a few comments and the setup methods for some tests.
