+++ src/main/java/interface_adaptors/servletLikePostGateway.java
@@ -1,6 +1,5 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
-import database_connection.Database;
 import database_connection.DatabaseUpdate;
 
 import javax.servlet.ServletException;
