+++ src/main/java/interface_adaptors/servletLoginGateway.java
@@ -1,6 +1,5 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
-import database_connection.Database;
 import database_connection.DatabaseRead;
 import log_in_use_case.LoginPasswordCheck;
 import user_exist_use_case.UserExistCheck;
