+++ src/main/java/interface_adaptors/servletPostPost.java
@@ -1,6 +1,4 @@
-package interfaceAdaptors;
-
-import database_connection.Database;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -9,10 +7,12 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletPostPostResponse class works as a presenter and pass message back to front-end.
-@WebServlet("/postPostResponse")
-public class servletPostPostResponse extends HttpServlet {
+// The servletPostPost class works as a controller and pass variables to gateway.
+@WebServlet("/listenPostPost")
+public class servletPostPost extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("success");
+        req.setCharacterEncoding("UTF-8");
+        req.getRequestDispatcher("/postPostGateway").forward(req,resp);
     }
 }
+
