+++ src/main/java/interface_adaptors/servletRegisterResponseUserExist.java
@@ -1,4 +1,4 @@
-package interfaceAdaptors;
+package interface_adaptors;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -7,11 +7,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-// The servletRegisterResponse class works as a presenter and pass success message back to front-end.
-@WebServlet("/registerResponse")
-public class servletRegisterResponse extends HttpServlet {
+// The servletRegisterResponseUserExist class works as a presenter and pass invalid message back to front-end.
+@WebServlet("/registerResponseUserExist")
+public class servletRegisterResponseUserExist extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        resp.getWriter().append("success");
+        resp.getWriter().append("username already exist");
     }
 }
-
