+++ src/test/java/common_user_test/CommonUserTest.java
@@ -1,6 +1,7 @@
 package common_user_test;
 
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import post_reply_user.CommonUser;
 import post_reply_user.Post;
@@ -11,18 +12,32 @@ import java.util.ArrayList;
 class CommonUserTest {
     //create some users, posts and replies
 
-    Post post1 = new Post("userId1", "post1 content");
+    Post post1 = null;
+    Post post2 = null;
+    Post post3 = null;
+    Post post4 = null;
 
-    Post post2 = new Post("userId1", "post2 content");
-    Post post3 = new Post("userId1", "post3 content");
-    Post post4 = new Post("userId1", "post4 content");
+    Reply reply1 = null;
+    Reply reply2 = null;
+    Reply reply3 = null;
+    Reply reply4 = null;
+    Reply reply5 = null;
 
-    Reply reply1 = new Reply("1","userId1","postId2","reply1","2022-12-01");
-    Reply reply2 = new Reply("2","userId2","postId3","reply2","2021-12-01");
-    Reply reply3 = new Reply("3","userId3","postId3","reply3","2020-12-01");
-    Reply reply4 = new Reply("4","userId4","postId3","reply4","2022-11-01");
-    Reply reply5 = new Reply("5","userId5","postId3","reply5","2021-11-01");
+    @BeforeEach
+    public void setup(){
 
+        post1 = new Post("userId1", "post1 content");
+        post2 = new Post("userId1", "post2 content");
+        post3 = new Post("userId1", "post3 content");
+        post4 = new Post("userId1", "post4 content");
+
+        reply1 = new Reply("1","userId1","postId2","reply1","2022-12-01");
+        reply2 = new Reply("2","userId2","postId3","reply2","2021-12-01");
+        reply3 = new Reply("3","userId3","postId3","reply3","2020-12-01");
+        reply4 = new Reply("4","userId4","postId3","reply4","2022-11-01");
+        reply5 = new Reply("5","userId5","postId3","reply5","2021-11-01");
+
+    }
 
     @Test
     public void testGetName() {
